pyproject.toml
[tool.poetry]
name = "bot-btc"
version = "0.1.0"
description = ""
authors = ["André Aquino <andre.luiz.aquino30@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
